refactor api sign-in and sign-up IR-411
