+++ src/main/java/com/cdcone/recipy/user/service/AuthService.java
@@ -4,24 +4,28 @@ import com.cdcone.recipy.user.dto.response.UserResponseDto;
 import com.cdcone.recipy.user.dto.request.SignInRequestDto;
 import com.cdcone.recipy.security.CustomUser;
 import com.cdcone.recipy.security.JwtUtil;
-import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-@Component
-@RequiredArgsConstructor
+@Service
 public class AuthService {
 
     private final AuthenticationManager authenticationManager;
     private final JwtUtil jwtUtil;
+
+    public AuthService(@Lazy AuthenticationManager authenticationManager, @Lazy JwtUtil jwtUtil) {
+        this.authenticationManager = authenticationManager;
+        this.jwtUtil = jwtUtil;
+    }
 
     public Optional<Map<String, Object>> auth(SignInRequestDto signInRequestDto) {
         String token;
